requirements.txt
linkify-it-py==2.0.3
markdown-it-py==3.0.0
mdformat==0.7.22
mdit-py-plugins==0.4.2
mdurl==0.1.2
uc-micro-py==1.0.3
